requirements.txt
openai==0.27.0
python-dotenv==0.21.1
PyQt6==6.5.2
PyPDF2==3.0.1
python-docx==0.8.11